# openapi.yaml
openapi: 3.0.3
info:
  title: Quesna Go API
  version: 2.0.0
  description: >
    Customer-app API for Quesna Go (Arabic + English).


    Auth: register or login -> OTP code is sent -> verify-otp -> returns a JWT token.
    In development mode the code is also returned as dev_otp.
    TEMP (until an SMS provider is connected): the fixed code 123456 is accepted for
    any account on verify-otp, and is returned as dev_otp on register/login/resend.
    Send the token on every protected call as `Authorization: Bearer <token>`.


    Language: pass `LANG: ar` or `LANG: en` as a request header on catalog / orders / notifications
    calls to get localized text. On register the UI language is the `?lang=` query parameter.


    Auth request bodies are form-data (JSON also accepted). Catalog/orders bodies are JSON.
    Every error carries a stable `error_code` the app localizes itself (see ERROR_CODES.md).


    This file covers the endpoints the customer app needs. Dashboard-only endpoints
    (/api/vendor, /api/admin, /api/dispatch, /api/driver) are intentionally not listed.
servers:
  - url: https://api.quesnago.com
    description: Production
  - url: http://localhost:4000
    description: Local
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register
      description: >
        Create an account and send an OTP code. Body is form-data (name, phone,
        village_id). The UI language is passed as the query parameter ?lang=ar|en
        (optional, default ar). Photo, email, birth_date and gender are added
        later via PATCH /api/auth/me.
      parameters:
        - in: query
          name: lang
          required: false
          schema: { type: string, enum: [ar, en], default: ar }
          description: "لغة واجهة التطبيق الحالية"
      requestBody:
        required: true
        content:
          multipart/form-data:
            example:
              name: "أحمد محمد"
              phone: "01065465118"
              village_id: 1
            schema:
              type: object
              required: [name, phone, village_id]
              properties:
                name: { type: string }
                phone:
                  type: string
                  description: Egyptian mobile. Accepts 01065465118, 1065465118 or +201065465118
                village_id: { type: integer }
      responses:
        '201':
          description: Account created, OTP sent
          content:
            application/json:
              example:
                success: true
                message: "تم إرسال كود التحقق"
                phone: "+201065465118"
                next: "otp"
                dev_otp: "483920"
              schema:
                $ref: '#/components/schemas/OtpSent'
        '409':
          description: Phone or email already used
          content:
            application/json:
              example: { success: false, error_code: "ALREADY_REGISTERED", error: "الرقم مسجّل بالفعل، من فضلك سجّل الدخول" }
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: Missing or invalid data
          content:
            application/json:
              example: { success: false, error_code: "INVALID_PHONE", error: "رقم موبايل مصري غير صحيح (مثال: 01065465118)" }
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Login
      description: Log in with the phone number and send an OTP code. Body is form-data.
      requestBody:
        required: true
        content:
          multipart/form-data:
            example:
              phone: "01065465118"
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              example:
                success: true
                message: "تم إرسال كود الدخول"
                phone: "+201065465118"
                next: "otp"
                dev_otp: "512804"
              schema:
                $ref: '#/components/schemas/OtpSent'
        '404':
          description: No account for this phone
          content:
            application/json:
              example: { success: false, error_code: "ACCOUNT_NOT_FOUND", error: "لا يوجد حساب بهذا الرقم، أنشئ حساب جديد" }
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: Invalid phone
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/verify-otp:
    post:
      tags: [Auth]
      operationId: verifyOtp
      summary: Verify OTP
      description: Confirm the OTP code and get the token + full user profile. Body is form-data.
      requestBody:
        required: true
        content:
          multipart/form-data:
            example:
              phone: "01065465118"
              code: "483920"
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string, description: 6 digits }
      responses:
        '200':
          description: Success, token returned
          content:
            application/json:
              example:
                success: true
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.payload.signature"
                user:
                  id: "3f2a9c10-7b1e-4d5a-9c2b-1e2f3a4b5c6d"
                  name: "أحمد محمد"
                  phone: "+201065465118"
                  email: "ahmed@example.com"
                  avatar_url: "https://i.pravatar.cc/150?img=12"
                  birth_date: "1998-04-25"
                  gender: "male"
                  lang: "ar"
                  village_id: 1
                  village_name: "كفر داود"
                  status: "active"
                  created_at: "2026-09-03T18:00:00.000Z"
                  is_new: true
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '400':
          description: Code expired or a new code is required
          content:
            application/json:
              example: { success: false, error_code: "OTP_EXPIRED", error: "انتهت صلاحية الكود، اطلب كود جديد" }
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Wrong code
          content:
            application/json:
              example: { success: false, error_code: "OTP_WRONG", error: "الكود غير صحيح (باقي 4 محاولات)", attempts_left: 4 }
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Too many attempts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/resend-otp:
    post:
      tags: [Auth]
      operationId: resendOtp
      summary: Resend OTP
      description: Send a new OTP code. Body is form-data.
      requestBody:
        required: true
        content:
          multipart/form-data:
            example:
              phone: "01065465118"
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string }
      responses:
        '200':
          description: New code sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpSent' }
        '404':
          description: No account for this phone
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/auth/me:
    get:
      tags: [Auth]
      operationId: currentUser
      summary: Current user
      description: Returns the full user profile. Requires a token.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              example:
                success: true
                user:
                  id: "3f2a9c10-7b1e-4d5a-9c2b-1e2f3a4b5c6d"
                  name: "أحمد محمد"
                  phone: "+201065465118"
                  email: "ahmed@example.com"
                  avatar_url: "https://i.pravatar.cc/150?img=12"
                  birth_date: "1998-04-25"
                  gender: "male"
                  lang: "ar"
                  village_id: 1
                  village_name: "كفر داود"
                  status: "active"
                  created_at: "2026-09-03T18:00:00.000Z"
                  updated_at: "2026-09-03T18:05:00.000Z"
                  phone_verified_at: "2026-09-03T18:00:30.000Z"
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  user: { $ref: '#/components/schemas/User' }
        '401':
          description: Login required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    patch:
      tags: [Auth]
      operationId: updateProfile
      summary: Update profile
      description: >
        Update the current user's profile. Requires a token. Body is form-data.
        All fields optional; send only what changes.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            example:
              name: "أحمد م."
              avatar_url: "https://i.pravatar.cc/150?img=5"
              email: "ahmed.new@example.com"
              birth_date: "1998-04-25"
              gender: "male"
              village_id: 2
              lang: "en"
            schema:
              type: object
              properties:
                name: { type: string }
                avatar_url: { type: string }
                email: { type: string }
                birth_date: { type: string, example: "1998-04-25" }
                gender: { type: string, enum: [male, female] }
                village_id: { type: integer }
                lang: { type: string, enum: [ar, en] }
      responses:
        '200':
          description: Updated, returns the full profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  user: { $ref: '#/components/schemas/User' }
        '401':
          description: Login required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Email already in use
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: Invalid data / nothing to update
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ============================ Catalog (public) ============================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OtpSent:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        phone: { type: string }
        next: { type: string }
        dev_otp:
          type: string
          description: >
            Present in development mode, or whenever the temporary fixed code is enabled
            (server env DEV_LOGIN_OTP). Currently 123456.
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        email: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
        birth_date: { type: string, nullable: true, example: "1998-04-25" }
        gender: { type: string, nullable: true, enum: [male, female, null] }
        lang: { type: string, enum: [ar, en] }
        village_id: { type: integer, nullable: true }
        village_name: { type: string, nullable: true }
        status: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        phone_verified_at: { type: string, format: date-time, nullable: true }
        is_new: { type: boolean, description: "true = أول تفعيل للحساب (على verify-otp فقط)" }
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        error_code:
          type: string
          description: كود ثابت يترجمه التطبيق (راجع ERROR_CODES.md)
          example: INVALID_PHONE
        error: { type: string, description: نص عربي للتشخيص فقط }
        message: { type: string, description: "نص الرسالة (بعض المسارات تستخدم message بدل error)" }
        attempts_left: { type: integer, description: "مع OTP_WRONG فقط" }
...